refactor(FR-1653): extract session creation logic into useStartSession hook (#4592)
# Refactor session creation logic into useStartSession hook

Resolves #4591 ([FR-1653](https://lablup.atlassian.net/browse/FR-1653))

## Summary

This PR extracts the complex session creation logic from the `SessionLauncherPage` component into a reusable custom hook called `useStartSession`. This refactoring improves code organization, maintainability, and testability.

## Changes

### New File
- **[react/src/hooks/useStartSession.tsx](react/src/hooks/useStartSession.tsx)** - Custom hook containing all session creation logic (~297 lines)
  - Handles session creation with all configuration options (batch, interactive, multiple sessions)
  - Manages error handling and notifications
  - Supports owner settings, resource allocation, and storage configuration

### Modified File
- **[react/src/pages/SessionLauncherPage.tsx](react/src/pages/SessionLauncherPage.tsx)** - Refactored to use the new hook
  - Removed ~280 lines of embedded session creation logic
  - Simplified component to focus on UI concerns
  - Maintained all existing functionality and behavior

## Benefits

1. **Better Separation of Concerns**: UI logic separated from business logic
2. **Improved Testability**: Session creation logic can now be tested in isolation
3. **Code Reusability**: Hook can be used in other components if needed
4. **Easier Maintenance**: Changes to session creation logic are now centralized
5. **Follows React Best Practices**: Proper use of custom hooks for complex logic

## Testing

All existing functionality has been preserved:
- Session creation with various configurations
- Batch mode and scheduling
- Resource allocation
- Storage mounting
- Environment variables
- Multiple session creation
- Error handling and notifications
- Navigation after creation

**Checklist:**

- [x] Code follows React best practices for custom hooks
- [x] All existing functionality is maintained
- [x] Error handling is preserved
- [x] No breaking changes to the API

[FR-1653]: https://lablup.atlassian.net/browse/FR-1653?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
